+++ R/cNORM.R
@@ -164,6 +164,8 @@ cNORM.GUI <- function(launch.browser=TRUE){
 #' (default R2 = 0.99)
 #' @param k The power constant. Higher values result in more detailed approximations
 #' but have the danger of over-fit (default = 4, max = 6)
+#' @param A the age power parameter (default NULL). If not set, cNORM automatically uses k. The age power parameter
+#' can be used to specify the k to produce rectangular matrices and specify the course of A independently from k
 #'
 #' @return cnorm object including the ranked raw data and the regression model
 #' @seealso rankByGroup, rankBySlidingWindow, computePowers, bestModel
@@ -214,6 +216,7 @@ cnorm <- function(raw = NULL,
                   method = 4,
                   descend = FALSE,
                   k = 4,
+                  A = NULL,
                   terms = 0,
                   R2 = NULL){
 
@@ -233,7 +236,7 @@ cnorm <- function(raw = NULL,
 
       message("Ranking data with sliding window ...")
       data <- rankBySlidingWindow(raw=raw, age=age, scale=scale, weights=weights, descend = descend, width = width, method = method)
-      data <- computePowers(data, k = k)
+      data <- computePowers(data, k = k, A = A)
     }else{
       data <- rankByGroup(raw=raw, group=group, scale=scale, weights=weights, descend = descend, method = method)
     }
@@ -243,10 +246,10 @@ cnorm <- function(raw = NULL,
         data <- computePowers(data, k = k)
       }else{
         data$age <- age
-        data <- computePowers(data, k = k, age = age)
+        data <- computePowers(data, k = k, A = A, age = age)
       }
     }else{
-      data <- computePowers(data, k = k)
+      data <- computePowers(data, k = k, A = A)
     }
   }
 
@@ -258,7 +261,7 @@ cnorm <- function(raw = NULL,
 
     message("Ranking data with sliding window ...")
     data <- rankBySlidingWindow(raw=raw, age=age, scale=scale, weights=weights, descend = descend, width = width, method = method)
-    data <- computePowers(data, k = k)
+    data <- computePowers(data, k = k, A = A)
   }else{
     stop("Please provide a numerical vector for the raw scores and either a vector for grouping and/or age of the same length. If you use an age vector only, please specify the width of the window.")
   }